+++ projects/core/src/lib/types/primitive.type.ts
@@ -1,1 +1,1 @@
-export type Primitive = string | number | bigint | boolean | symbol | null | undefined;
+export type IPrimitive = string | number | bigint | boolean | symbol | null | undefined;