+++ projects/ui-kit/src/lib/ui-kit/table/util/default-pagination.ts
@@ -1,8 +1,8 @@
-import { PageModel } from '@rt-tools/utils';
+import { IPageModel } from '@rt-tools/utils';
 
 export const DEFAULT_PAGE_SIZE: number = 20;
 
-export const DEFAULT_PAGE_MODEL: Readonly<PageModel> = Object.freeze({
+export const DEFAULT_PAGE_MODEL: Readonly<IPageModel> = Object.freeze({
     pageNumber: 1,
     pageSize: DEFAULT_PAGE_SIZE,
     totalCount: 0,