+++ projects/ui-kit/src/lib/ui-kit/toolbar/toolbar.component.ts
@@ -14,7 +14,7 @@ import {
 import { MatToolbar } from '@angular/material/toolbar';
 
 import { BlockDirective, ElemDirective, ModDirective } from '@rt-tools/core';
-import { Nullable } from '@rt-tools/utils';
+import { INullable } from '@rt-tools/utils';
 
 @Directive({
     selector: '[rtuiToolbarLeft]',
@@ -41,16 +41,16 @@ export class RtuiToolbarRightDirective {}
 export class RtuiToolbarComponent {
     public isVisibleToolbar: Signal<boolean> = input(true);
 
-    public readonly leftToolTpl: Signal<Nullable<TemplateRef<Type<unknown>>>> = contentChild(RtuiToolbarLeftDirective, {
+    public readonly leftToolTpl: Signal<INullable<TemplateRef<Type<unknown>>>> = contentChild(RtuiToolbarLeftDirective, {
         read: TemplateRef,
     });
-    public readonly centerToolTpl: Signal<Nullable<TemplateRef<Type<unknown>>>> = contentChild(RtuiToolbarCenterDirective, {
+    public readonly centerToolTpl: Signal<INullable<TemplateRef<Type<unknown>>>> = contentChild(RtuiToolbarCenterDirective, {
         read: TemplateRef,
     });
-    public readonly rightToolTpl: Signal<Nullable<TemplateRef<Type<unknown>>>> = contentChild(RtuiToolbarRightDirective, {
+    public readonly rightToolTpl: Signal<INullable<TemplateRef<Type<unknown>>>> = contentChild(RtuiToolbarRightDirective, {
         read: TemplateRef,
     });
-    public sticky: InputSignalWithTransform<Nullable<boolean>, boolean> = input<Nullable<boolean>, boolean>(false, {
+    public sticky: InputSignalWithTransform<INullable<boolean>, boolean> = input<INullable<boolean>, boolean>(false, {
         transform: booleanAttribute,
     });
 }