+++ projects/utils/src/lib/interfaces/basic.types.ts
@@ -1,3 +1,3 @@
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type AnyObject = Record<string, any>;
-export type EmptyObject = Record<string, never>;
+export type IAnyObject = Record<string, any>;
+export type IEmptyObject = Record<string, never>;