+++ projects/utils/src/lib/interfaces/intersection.type.ts
@@ -1,1 +1,1 @@
-export type IntersectionType<T extends object, M extends object> = keyof T & keyof M;
+export type IIntersectionType<T extends object, M extends object> = keyof T & keyof M;